+++ apimocking-main/apimocking-boilerplate/src/main/java/com/grepp/spring/app/model/budget/model/BudgetCalenderResponseDto.java
@@ -1,5 +1,6 @@
 package com.grepp.spring.app.model.budget.model;
 
+import java.math.BigDecimal;
 import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -14,8 +15,8 @@ import lombok.Setter;
 public class BudgetCalenderResponseDto {
 
     private String month;
-    private int totalIncome;
-    private int totalExpense;
+    private BigDecimal totalIncome;
+    private BigDecimal totalExpense;
     private List<BudgetDaySummary> days;
 
 }
